refactor: move logging api to core
